Adding a build step to test build of home directory.
Various fixes in the system and home directory roles to make this possible.

First tried also making a Docker image of a system with the home directory set up, but ran out of space. Will instead proceed to test the different roles separately, and use the docker images only as test platforms.